removed all LZ legacy stuff

+++ contracts/Game.sol
@@ -37,13 +37,13 @@ contract Game is ERC721, ReentrancyGuard {
     // rebalance period of vault, upped at vault rebalance
     uint256 rebalancingPeriod;
     // (chainId => vaultAddress)
-    mapping(uint16 => address) vaultAddress;
+    mapping(uint32 => address) vaultAddress;
     // (chainId => deltaAllocation)
     mapping(uint256 => int256) deltaAllocationChain;
     // (chainId => protocolNumber => deltaAllocation)
     mapping(uint256 => mapping(uint256 => int256)) deltaAllocationProtocol;
     // (chainId => rebalancing period => protocol id => rewardPerLockedToken).
-    mapping(uint16 => mapping(uint256 => mapping(uint256 => int256))) rewardPerLockedToken;
+    mapping(uint32 => mapping(uint256 => mapping(uint256 => int256))) rewardPerLockedToken;
   }
 
   address private dao;
@@ -58,7 +58,7 @@ contract Game is ERC721, ReentrancyGuard {
   uint256 private latestBasketId;
 
   // array of chainIds e.g [10, 100, 1000];
-  uint16[] public chainIds;
+  uint32[] public chainIds;
 
   // interval in Unix timeStamp
   uint256 public rebalanceInterval; // SHOULD BE REPLACED FOR REALISTIC NUMBER
@@ -84,7 +84,7 @@ contract Game is ERC721, ReentrancyGuard {
   // (vaultNumber => bool): true when vault is cross-chain rebalancing
   mapping(uint256 => bool) public isXChainRebalancing;
 
-  event PushProtocolAllocations(uint16 chain, address vault, int256[] deltas);
+  event PushProtocolAllocations(uint32 chain, address vault, int256[] deltas);
 
   event PushedAllocationsToController(uint256 vaultNumber, int256[] deltas);
 
@@ -346,7 +346,7 @@ contract Game is ERC721, ReentrancyGuard {
   ) internal returns (int256 totalDelta) {
     for (uint256 i = 0; i < _deltaAllocations.length; i++) {
       int256 chainTotal;
-      uint16 chain = chainIds[i];
+      uint32 chain = chainIds[i];
       uint256 latestProtocol = latestProtocolId[chain];
       require(_deltaAllocations[i].length == latestProtocol, "Invalid allocation length");
 
@@ -375,7 +375,7 @@ contract Game is ERC721, ReentrancyGuard {
     if (currentRebalancingPeriod <= lastRebalancingPeriod) return;
 
     for (uint k = 0; k < chainIds.length; k++) {
-      uint16 chain = chainIds[k];
+      uint32 chain = chainIds[k];
       uint256 latestProtocol = latestProtocolId[chain];
       for (uint i = 0; i < latestProtocol; i++) {
         int256 allocation = basketAllocationInProtocol(_basketId, chain, i) / 1E18;
@@ -441,7 +441,7 @@ contract Game is ERC721, ReentrancyGuard {
     deltas = new int[](chainIds.length);
 
     for (uint256 i = 0; i < chainIds.length; i++) {
-      uint16 chain = chainIds[i];
+      uint32 chain = chainIds[i];
       deltas[i] = getDeltaAllocationChain(_vaultNumber, chain);
       vaults[_vaultNumber].deltaAllocationChain[chain] = 0;
     }
@@ -455,7 +455,7 @@ contract Game is ERC721, ReentrancyGuard {
     require(isXChainRebalancing[_vaultNumber], "Vault is not rebalancing");
 
     for (uint256 i = 0; i < chainIds.length; i++) {
-      uint16 chain = chainIds[i];
+      uint32 chain = chainIds[i];
       int256[] memory deltas = protocolAllocationsToArray(_vaultNumber, chain);
 
       IXProvider(xProvider).pushProtocolAllocationsToVault{value: msg.value}(
@@ -476,7 +476,7 @@ contract Game is ERC721, ReentrancyGuard {
   /// @return deltas Array with allocations where the index matches the protocolId
   function protocolAllocationsToArray(
     uint256 _vaultNumber,
-    uint16 _chainId
+    uint32 _chainId
   ) internal returns (int256[] memory deltas) {
     uint256 latestId = latestProtocolId[_chainId];
     deltas = new int[](latestId);
@@ -490,7 +490,7 @@ contract Game is ERC721, ReentrancyGuard {
   /// @notice See settleRewardsInt below
   function settleRewards(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     int256[] memory _rewards
   ) external onlyXProvider {
     settleRewardsInt(_vaultNumber, _chainId, _rewards);
@@ -504,7 +504,7 @@ contract Game is ERC721, ReentrancyGuard {
   /// @param _rewards Array with rewardsPerLockedToken of all protocols in vault => index matches protocolId
   function settleRewardsInt(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     int256[] memory _rewards
   ) internal {
     uint256 rebalancingPeriod = vaults[_vaultNumber].rebalancingPeriod;
@@ -525,7 +525,7 @@ contract Game is ERC721, ReentrancyGuard {
   /// @notice Getter for rewardsPerLockedToken for given vaultNumber => chainId => rebalancingPeriod => protocolId
   function getRewardsPerLockedToken(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     uint256 _rebalancingPeriod,
     uint256 _protocolId
   ) internal view returns (int256) {
@@ -552,7 +552,7 @@ contract Game is ERC721, ReentrancyGuard {
   }
 
   /// @notice getter for vault address linked to a chainId
-  function getVaultAddress(uint256 _vaultNumber, uint16 _chainId) internal view returns (address) {
+  function getVaultAddress(uint256 _vaultNumber, uint32 _chainId) internal view returns (address) {
     return vaults[_vaultNumber].vaultAddress[_chainId];
   }
 
@@ -567,7 +567,7 @@ contract Game is ERC721, ReentrancyGuard {
   }
 
   /// @notice Getter for chainId array
-  function getChainIds() public view returns (uint16[] memory) {
+  function getChainIds() public view returns (uint32[] memory) {
     return chainIds;
   }
 
@@ -630,7 +630,7 @@ contract Game is ERC721, ReentrancyGuard {
   /// @notice setter to link a chainId to a vault address for cross chain functions
   function setVaultAddress(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     address _address
   ) external onlyGuardian {
     vaults[_vaultNumber].vaultAddress[_chainId] = _address;
@@ -639,13 +639,13 @@ contract Game is ERC721, ReentrancyGuard {
   /// @notice Setter for latest protocol Id for given chainId.
   /// @param _chainId number of chain id set in chainIds array
   /// @param _latestProtocolId latest protocol Id aka number of supported protocol vaults, starts at 0
-  function setLatestProtocolId(uint16 _chainId, uint256 _latestProtocolId) external onlyGuardian {
+  function setLatestProtocolId(uint32 _chainId, uint256 _latestProtocolId) external onlyGuardian {
     latestProtocolId[_chainId] = _latestProtocolId;
   }
 
   /// @notice Setter for chainId array
   /// @param _chainIds array of all the used chainIds
-  function setChainIds(uint16[] memory _chainIds) external onlyGuardian {
+  function setChainIds(uint32[] memory _chainIds) external onlyGuardian {
     chainIds = _chainIds;
   }
 
@@ -657,7 +657,7 @@ contract Game is ERC721, ReentrancyGuard {
   /// @notice Step 8: Guardian function
   function settleRewardsGuard(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     int256[] memory _rewards
   ) external onlyGuardian {
     settleRewardsInt(_vaultNumber, _chainId, _rewards);
