+++ contracts/XChainController.sol
@@ -18,19 +18,19 @@ contract XChainController {
     uint256 totalSupply;
     uint256 totalWithdrawalRequests;
     // (chainId => bool): true == off // false == on
-    mapping(uint16 => bool) chainIdOff;
+    mapping(uint32 => bool) chainIdOff;
     // (chainId => currentAllocation)
-    mapping(uint16 => int256) currentAllocationPerChain;
+    mapping(uint32 => int256) currentAllocationPerChain;
     // (chainId => totalUnderlying)
-    mapping(uint16 => uint256) totalUnderlyingPerChain;
+    mapping(uint32 => uint256) totalUnderlyingPerChain;
     // (chainId => vaultAddress)
-    mapping(uint16 => address) vaultChainAddress;
+    mapping(uint32 => address) vaultChainAddress;
     // (chainId => underlyingAddress): e.g USDC
-    mapping(uint16 => address) vaultUnderlyingAddress;
+    mapping(uint32 => address) vaultUnderlyingAddress;
     // (chainId => totalWithdrawalRequests): total withdrawal requests in LP Token
-    mapping(uint16 => uint256) withdrawalRequests;
+    mapping(uint32 => uint256) withdrawalRequests;
     // (chainId => amountToDeposit)
-    mapping(uint16 => uint256) amountToDepositPerChain;
+    mapping(uint32 => uint256) amountToDepositPerChain;
   }
 
   // activeVaults; number of active vaults for vaultNumber, set in XChainRebalance
@@ -52,8 +52,8 @@ contract XChainController {
   address public xProviderAddr;
   IXProvider public xProvider;
 
-  uint16[] public chainIds;
-  uint16 public homeChain;
+  uint32[] public chainIds;
+  uint32 public homeChain;
 
   // (vaultNumber => vaultInfo struct)
   mapping(uint256 => vaultInfo) internal vaults;
@@ -62,12 +62,12 @@ contract XChainController {
 
   event SendXChainAmount(
     address _vault,
-    uint16 _chainId,
+    uint32 _chainId,
     uint256 _amountToSendXChain,
     uint256 _exchangeRate
   );
 
-  event SentFundsToVault(address _vault, uint16 _chainId, uint256 _amount, address _asset);
+  event SentFundsToVault(address _vault, uint32 _chainId, uint256 _amount, address _asset);
 
   modifier onlyGame() {
     require(msg.sender == game, "xController: only Game");
@@ -119,7 +119,7 @@ contract XChainController {
     _;
   }
 
-  constructor(address _game, address _dao, address _guardian, uint16 _homeChain) {
+  constructor(address _game, address _dao, address _guardian, uint32 _homeChain) {
     game = _game;
     dao = _dao;
     guardian = _guardian;
@@ -180,7 +180,7 @@ contract XChainController {
   }
 
   /// @notice Resets underlying for a vaultNumber per chainId at the start of a rebalancing period
-  function resetVaultUnderlyingForChain(uint256 _vaultNumber, uint16 _chainId) internal {
+  function resetVaultUnderlyingForChain(uint256 _vaultNumber, uint32 _chainId) internal {
     vaults[_vaultNumber].totalUnderlyingPerChain[_chainId] = 0;
   }
 
@@ -201,7 +201,7 @@ contract XChainController {
     uint256 activeVaults;
 
     for (uint256 i = 0; i < chainIds.length; i++) {
-      uint16 chain = chainIds[i];
+      uint32 chain = chainIds[i];
       activeVaults += settleCurrentAllocation(_vaultNumber, chain, _deltas[i]);
       resetVaultUnderlyingForChain(_vaultNumber, chain);
     }
@@ -220,7 +220,7 @@ contract XChainController {
   /// @param _deltas Delta allocations array received from game, indexes match chainIds[] set in this contract
   function settleCurrentAllocation(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     int256 _deltas
   ) internal returns (uint256 activeVault) {
     if (getCurrentAllocation(_vaultNumber, _chainId) == 0 && _deltas == 0) {
@@ -239,7 +239,7 @@ contract XChainController {
 
   /// @notice Will send feedback to the vault if it is turned on or off by settleCurrentAllocation
   /// @param _state Bool if vault is turned on or off
-  function sendFeedbackToVault(uint256 _vaultNumber, uint16 _chainId, bool _state) internal {
+  function sendFeedbackToVault(uint256 _vaultNumber, uint32 _chainId, bool _state) internal {
     if (getVaultChainIdOff(_vaultNumber, _chainId) != _state) {
       xProvider.pushStateFeedbackToVault{value: msg.value}(
         getVaultAddress(_vaultNumber, _chainId),
@@ -254,7 +254,7 @@ contract XChainController {
   /// @notice See setTotalUnderlyingInt below
   function setTotalUnderlying(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     uint256 _underlying,
     uint256 _totalSupply,
     uint256 _withdrawalRequests
@@ -272,7 +272,7 @@ contract XChainController {
   /// @param _withdrawalRequests Total amount of withdrawal requests from the vault in LP Tokens
   function setTotalUnderlyingInt(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     uint256 _underlying,
     uint256 _totalSupply,
     uint256 _withdrawalRequests
@@ -300,7 +300,7 @@ contract XChainController {
     uint256 newExchangeRate = (totalUnderlying * (10 ** decimals)) / totalSupply;
 
     for (uint i = 0; i < chainIds.length; i++) {
-      uint16 chain = chainIds[i];
+      uint32 chain = chainIds[i];
       if (getVaultChainIdOff(_vaultNumber, chain)) continue;
 
       int256 amountToChain = calcAmountToChain(
@@ -325,7 +325,7 @@ contract XChainController {
   /// @param _amountToChain Amount in vaultcurrency that should be on given chainId
   function calcDepositWithdraw(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     int256 _amountToChain
   ) internal view returns (int256, uint256) {
     uint256 currentUnderlying = getTotalUnderlyingOnChain(_vaultNumber, _chainId);
@@ -343,7 +343,7 @@ contract XChainController {
   /// @param _totalAllocation Total allocation on all chains for given vaultNumber
   function calcAmountToChain(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     uint256 _totalUnderlying,
     int256 _totalAllocation
   ) internal view returns (int256) {
@@ -363,7 +363,7 @@ contract XChainController {
   /// @param _exchangeRate New exchangerate for vaults
   function sendXChainAmount(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     int256 _amountDeposit,
     uint256 _amountToWithdraw,
     uint256 _exchangeRate
@@ -396,7 +396,7 @@ contract XChainController {
     uint256 _slippage
   ) external payable onlyWhenFundsReceived(_vaultNumber) {
     for (uint i = 0; i < chainIds.length; i++) {
-      uint16 chain = chainIds[i];
+      uint32 chain = chainIds[i];
       if (getVaultChainIdOff(_vaultNumber, chain)) continue;
 
       uint256 amountToDeposit = getAmountToDeposit(_vaultNumber, chain);
@@ -409,7 +409,13 @@ contract XChainController {
         if (amountToDeposit > balance) amountToDeposit = balance;
 
         IERC20(underlying).safeIncreaseAllowance(address(xProvider), amountToDeposit);
-        xProvider.xTransferToVaults{value: msg.value}(vault, chain, amountToDeposit, underlying, _slippage);
+        xProvider.xTransferToVaults{value: msg.value}(
+          vault,
+          chain,
+          amountToDeposit,
+          underlying,
+          _slippage
+        );
         setAmountToDeposit(_vaultNumber, chain, 0);
         emit SentFundsToVault(vault, chain, amountToDeposit, underlying);
       }
@@ -421,7 +427,7 @@ contract XChainController {
   /// @notice Helper to get total current allocation of vaultNumber
   function getTotalUnderlyingOnChain(
     uint256 _vaultNumber,
-    uint16 _chainId
+    uint32 _chainId
   ) internal view returns (uint256) {
     return vaults[_vaultNumber].totalUnderlyingPerChain[_chainId];
   }
@@ -434,14 +440,14 @@ contract XChainController {
   }
 
   /// @notice Helper to get vault address of vaultNumber with given chainID
-  function getVaultAddress(uint256 _vaultNumber, uint16 _chainId) internal view returns (address) {
+  function getVaultAddress(uint256 _vaultNumber, uint32 _chainId) internal view returns (address) {
     return vaults[_vaultNumber].vaultChainAddress[_chainId];
   }
 
   /// @notice Helper to get underyling address of vaultNumber with given chainID eg USDC
   function getUnderlyingAddress(
     uint256 _vaultNumber,
-    uint16 _chainId
+    uint32 _chainId
   ) internal view returns (address) {
     return vaults[_vaultNumber].vaultUnderlyingAddress[_chainId];
   }
@@ -449,7 +455,7 @@ contract XChainController {
   /// @notice Helper to get current allocation per chain of vaultNumber with given chainID
   function getCurrentAllocation(
     uint256 _vaultNumber,
-    uint16 _chainId
+    uint32 _chainId
   ) internal view returns (int256) {
     return vaults[_vaultNumber].currentAllocationPerChain[_chainId];
   }
@@ -460,14 +466,14 @@ contract XChainController {
   }
 
   /// @notice Helper to get if vault is active or not
-  function getVaultChainIdOff(uint256 _vaultNumber, uint16 _chainId) public view returns (bool) {
+  function getVaultChainIdOff(uint256 _vaultNumber, uint32 _chainId) public view returns (bool) {
     return vaults[_vaultNumber].chainIdOff[_chainId];
   }
 
   /// @notice Helper to set the amount to deposit in a chain vault
   function setAmountToDeposit(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     int256 _amountToDeposit
   ) internal {
     vaults[_vaultNumber].amountToDepositPerChain[_chainId] = uint256(_amountToDeposit);
@@ -476,7 +482,7 @@ contract XChainController {
   /// @notice Helper to get the amount to deposit in a chain vault
   function getAmountToDeposit(
     uint256 _vaultNumber,
-    uint16 _chainId
+    uint32 _chainId
   ) internal view returns (uint256) {
     return vaults[_vaultNumber].amountToDepositPerChain[_chainId];
   }
@@ -489,7 +495,7 @@ contract XChainController {
   /// @notice Helper to get withdrawal requests from the vault on given chainId
   function getWithdrawalRequests(
     uint256 _vaultNumber,
-    uint16 _chainId
+    uint32 _chainId
   ) internal view returns (uint256) {
     return vaults[_vaultNumber].withdrawalRequests[_chainId];
   }
@@ -500,7 +506,7 @@ contract XChainController {
   }
 
   /// @notice Getter for chainId array
-  function getChainIds() public view returns (uint16[] memory) {
+  function getChainIds() public view returns (uint32[] memory) {
     return chainIds;
   }
 
@@ -525,7 +531,7 @@ contract XChainController {
   /// @param _underlying underlying of the Vault eg USDC
   function setVaultChainAddress(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     address _address,
     address _underlying
   ) external onlyDao {
@@ -541,7 +547,7 @@ contract XChainController {
 
   /// @notice Setter for homeChain Id
   /// @param _homeChainId New home chainId
-  function setHomeChainId(uint16 _homeChainId) external onlyDao {
+  function setHomeChainId(uint32 _homeChainId) external onlyDao {
     homeChain = _homeChainId;
   }
 
@@ -569,7 +575,7 @@ contract XChainController {
 
   /// @notice Setter for chainId array
   /// @param _chainIds array of all the used chainIds
-  function setChainIds(uint16[] memory _chainIds) external onlyGuardian {
+  function setChainIds(uint32[] memory _chainIds) external onlyGuardian {
     chainIds = _chainIds;
   }
 
@@ -591,7 +597,7 @@ contract XChainController {
   /// @notice Step 2: Guardian function
   function setTotalUnderlyingGuard(
     uint256 _vaultNumber,
-    uint16 _chainId,
+    uint32 _chainId,
     uint256 _underlying,
     uint256 _totalSupply,
     uint256 _withdrawalRequests
